Recurring categories (vigência, editable base, inferred credit/debit),
anti-duplication, keyword rules over card + extrato, Categorias page,
Despesas & Receitas redesign, money masks, Nielsen skill. v0.2.1.

+++ src-tauri/src/domain/categorizer.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 pub struct CategoryRule {
     pub keywords: Vec<String>,
     pub category: String,
-    pub priority: u8,
+    pub priority: u32,
 }
 
 pub struct Categorizer {
